+++ NEWS.md
@@ -1,4 +1,4 @@
-# resourcecode (development version)
+# resourcecode 0.5.0
 
 - Update the data retrieving function to fail gracefully
 whenever the remote database is not accessible.